+++ src/GafferScene/SceneAlgo.cpp
@@ -37,11 +37,15 @@
 #include "GafferScene/SceneAlgo.h"
 
 #include "GafferScene/CameraTweaks.h"
+#include "GafferScene/CopyAttributes.h"
 #include "GafferScene/Filter.h"
 #include "GafferScene/FilterProcessor.h"
+#include "GafferScene/LocaliseAttributes.h"
+#include "GafferScene/MergeScenes.h"
 #include "GafferScene/PathFilter.h"
 #include "GafferScene/ScenePlug.h"
 #include "GafferScene/ShaderTweaks.h"
+#include "GafferScene/ShuffleAttributes.h"
 
 #include "Gaffer/Context.h"
 #include "Gaffer/Monitor.h"
@@ -423,6 +427,125 @@ int filterResult( const FilterPlug *filter, const ScenePlug *scene )
 	return filter->getValue();
 }
 
+void addGenericAttributePredecessors( const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
+{
+	for( auto &h : source )
+	{
+		if( auto ah = SceneAlgo::attributeHistory( h.get(), destination->attributeName ) )
+		{
+			destination->predecessors.push_back( ah );
+		}
+	}
+}
+
+void addCopyAttributesPredecessors( const CopyAttributes *copyAttributes, const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
+{
+	const ScenePlug *sourceScene = copyAttributes->inPlug();
+	if(
+		( filterResult( copyAttributes->filterPlug(), copyAttributes->inPlug() ) & PathMatcher::ExactMatch ) &&
+		StringAlgo::matchMultiple( destination->attributeName, copyAttributes->attributesPlug()->getValue() )
+	)
+	{
+		ConstCompoundObjectPtr sourceAttributes;
+		const std::string sourceLocation = copyAttributes->sourceLocationPlug()->getValue();
+		if( sourceLocation.empty() )
+		{
+			if( copyAttributes->sourcePlug()->exists() )
+			{
+				sourceAttributes = copyAttributes->sourcePlug()->attributesPlug()->getValue();
+			}
+		}
+		else
+		{
+			ScenePlug::ScenePath sourcePath; ScenePlug::stringToPath( sourceLocation, sourcePath );
+			if( copyAttributes->sourcePlug()->exists( sourcePath ) )
+			{
+				sourceAttributes = copyAttributes->sourcePlug()->attributes( sourcePath );
+			}
+		}
+		if( sourceAttributes && sourceAttributes->members().count( destination->attributeName ) )
+		{
+			sourceScene = copyAttributes->sourcePlug();
+		}
+	}
+
+	for( auto &h : source )
+	{
+		if( h->scene == sourceScene )
+		{
+			destination->predecessors.push_back( SceneAlgo::attributeHistory( h.get(), destination->attributeName ) );
+		}
+	}
+}
+
+void addShuffleAttributesPredecessors( const ShuffleAttributes *shuffleAttributes, const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
+{
+	// We have no way of introspecting the operation of a ShufflePlug, so we resort
+	// to shuffling	`name = name, value = name` pairs to figure out where the attribute
+	// has come from.
+
+	InternedString sourceAttributeName = destination->attributeName;
+	if( filterResult( shuffleAttributes->filterPlug(), shuffleAttributes->inPlug() ) & PathMatcher::ExactMatch )
+	{
+		auto inputAttributes = shuffleAttributes->inPlug()->attributesPlug()->getValue();
+		map<InternedString, InternedString> shuffledNames;
+		for( auto &a : inputAttributes->members() )
+		{
+			shuffledNames.insert( { a.first, a.first } );
+		}
+		shuffledNames = shuffleAttributes->shufflesPlug()->shuffle( shuffledNames );
+		sourceAttributeName = shuffledNames[destination->attributeName];
+	}
+
+	assert( source.size() == 1 );
+	destination->predecessors.push_back( SceneAlgo::attributeHistory( source[0].get(), sourceAttributeName ) );
+}
+
+void addLocaliseAttributesPredecessors( const LocaliseAttributes *localiseAttributes, const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
+{
+	// No need to check if the node is filtered to this location.
+	// Filtered or unfiltered, it's all the same : the predecessor
+	// we want is the most local one. i.e. the one with the longest
+	// path.
+
+	int longestPath = -1;
+	SceneAlgo::AttributeHistory::Ptr predecessor;
+	for( auto &h : source )
+	{
+		const auto &sourcePath = h->context->get<ScenePlug::ScenePath>( ScenePlug::scenePathContextName );
+		if( (int)sourcePath.size() <= longestPath )
+		{
+			continue;
+		}
+		if( auto p = attributeHistory( h.get(), destination->attributeName ) )
+		{
+			predecessor = p;
+			longestPath = sourcePath.size();
+		}
+	}
+
+	assert( predecessor );
+	destination->predecessors.push_back( predecessor );
+}
+
+void addMergeScenesPredecessors( const MergeScenes *mergeScenes, const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
+{
+	// MergeScenes only evaluates input locations that exist, and in an order
+	// whereby the last input with the attribute wins.
+
+	SceneAlgo::AttributeHistory::Ptr predecessor;
+	for( auto &h : source )
+	{
+		if( auto p = attributeHistory( h.get(), destination->attributeName ) )
+		{
+			predecessor = p;
+		}
+	}
+
+	assert( predecessor );
+	destination->predecessors.push_back( predecessor );
+}
+
 SceneProcessor *objectTweaksWalk( const SceneAlgo::History *h )
 {
 	if( auto tweaks = h->scene->parent<CameraTweaks>() )
@@ -448,7 +571,7 @@ SceneProcessor *objectTweaksWalk( const SceneAlgo::History *h )
 	return nullptr;
 }
 
-ShaderTweaks *shaderTweaksWalk( const SceneAlgo::History *h, const IECore::InternedString &attributeName )
+ShaderTweaks *shaderTweaksWalk( const SceneAlgo::AttributeHistory *h )
 {
 	if( auto tweaks = h->scene->parent<ShaderTweaks>() )
 	{
@@ -456,7 +579,7 @@ ShaderTweaks *shaderTweaksWalk( const SceneAlgo::History *h, const IECore::Inter
 		{
 			Context::Scope contextScope( h->context.get() );
 			if(
-				StringAlgo::matchMultiple( attributeName, tweaks->shaderPlug()->getValue() ) &&
+				StringAlgo::matchMultiple( h->attributeName, tweaks->shaderPlug()->getValue() ) &&
 				( filterResult( tweaks->filterPlug(), tweaks->inPlug() ) & PathMatcher::ExactMatch )
 			)
 			{
@@ -467,7 +590,7 @@ ShaderTweaks *shaderTweaksWalk( const SceneAlgo::History *h, const IECore::Inter
 
 	for( const auto &p : h->predecessors )
 	{
-		if( auto tweaks = shaderTweaksWalk( p.get(), attributeName ) )
+		if( auto tweaks = shaderTweaksWalk( static_cast<SceneAlgo::AttributeHistory *>( p.get() ) ) )
 		{
 			return tweaks;
 		}
@@ -508,6 +631,61 @@ SceneAlgo::History::Ptr SceneAlgo::history( const Gaffer::ValuePlug *scenePlugCh
 	return historyWalk( monitor->rootProcesses().front().get(), scenePlugChild->getName(), nullptr );
 }
 
+SceneAlgo::AttributeHistory::Ptr SceneAlgo::attributeHistory( const SceneAlgo::History *attributesHistory, const IECore::InternedString &attribute )
+{
+	Context::Scope scopedContext( attributesHistory->context.get() );
+	ConstCompoundObjectPtr attributes = attributesHistory->scene->attributesPlug()->getValue();
+	ConstObjectPtr attributeValue = attributes->member<Object>( attribute );
+
+	if( !attributeValue )
+	{
+		return nullptr;
+	}
+
+	SceneAlgo::AttributeHistory::Ptr result = new AttributeHistory(
+		attributesHistory->scene, attributesHistory->context,
+		attribute, attributeValue
+	);
+
+	// Filter the _attributes_ history to include only predecessors which
+	// contribute specifically to our single _attribute_. In the absence of
+	// a SceneNode-level API for querying attribute sources, we resort to
+	// special case code for backtracking through certain node types.
+	/// \todo Consider an official API that allows the nodes themselves to
+	/// take responsibility for this backtracking.
+
+	auto node = runTimeCast<const SceneNode>( attributesHistory->scene->node() );
+	if( node && node->enabledPlug()->getValue() && attributesHistory->scene == node->outPlug() )
+	{
+		if( auto copyAttributes = runTimeCast<const CopyAttributes>( node ) )
+		{
+			addCopyAttributesPredecessors( copyAttributes, attributesHistory->predecessors, result.get() );
+		}
+		else if( auto shuffleAttributes = runTimeCast<const ShuffleAttributes>( node ) )
+		{
+			addShuffleAttributesPredecessors( shuffleAttributes, attributesHistory->predecessors, result.get() );
+		}
+		else if( auto localiseAttributes = runTimeCast<const LocaliseAttributes>( node ) )
+		{
+			addLocaliseAttributesPredecessors( localiseAttributes, attributesHistory->predecessors, result.get() );
+		}
+		else if( auto mergeScenes = runTimeCast<const MergeScenes>( node ) )
+		{
+			addMergeScenesPredecessors( mergeScenes, attributesHistory->predecessors, result.get() );
+		}
+		else
+		{
+			addGenericAttributePredecessors( attributesHistory->predecessors, result.get() );
+		}
+	}
+	else
+	{
+		addGenericAttributePredecessors( attributesHistory->predecessors, result.get() );
+	}
+
+	return result;
+}
+
 ScenePlug *SceneAlgo::source( const ScenePlug *scene, const ScenePlug::ScenePath &path )
 {
 	History::ConstPtr h = history( scene->objectPlug(), path );
@@ -544,14 +722,10 @@ ShaderTweaks *SceneAlgo::shaderTweaks( const ScenePlug *scene, const ScenePlug::
 	ScenePlug::ScenePath inheritancePath = path;
 	while( inheritancePath.size() )
 	{
-		ConstCompoundObjectPtr attributes = scene->attributes( inheritancePath );
-		if( attributes->member<Object>( attributeName ) )
+		History::ConstPtr h = history( scene->attributesPlug(), inheritancePath );
+		if( auto ah = attributeHistory( h.get(), attributeName ) )
 		{
-			History::ConstPtr h = history( scene->attributesPlug(), inheritancePath );
-			if( h )
-			{
-				return shaderTweaksWalk( h.get(), attributeName );
-			}
+			return shaderTweaksWalk( ah.get() );
 		}
 		inheritancePath.pop_back();
 	}